requirements.txt
torch
torchvision
timm
Pillow
pyyaml

gradio==4.44.0
huggingface_hub==0.23.5